renamed hmcecs_utils to ecs_utils and added todos.
